refactor: extract controller settings form into its own component
